m8: disable libaudioamp for now
Fixes the "buzzing sound bug". The bug appears after a phone call
when volume is turned up to 80% or higher.
Users have found out that the audio amplifier is responsible for
that, so lets leave it out until we have a proper fix for that.

Removing the amplifier doesn't have any noticeable disadvantages.

Change-Id: Id2bb8e7ea31404bc89705470820d000d2e8638aa
Signed-off-by: noname55804 <marlin2201@gmail.com>